ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
pypdf2==3.0.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
llama-index==0.9.6
llama-index-llms-openai==0.1.5
llama-index-embeddings-openai==0.1.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2